+++ inc/functions.php
@@ -1,7 +1,7 @@
 <?php
 /**
- * @package WPOD
- * @version 0.1.1
+ * @package WPUpstream
+ * @version 0.1.2
  * @author Usability Dynamics Inc.
  */
 
@@ -15,7 +15,6 @@ function wpupstream_display_wpversion_error_notice() {
 	<div class="error">
 		<p><?php printf( __( 'Fatal error with plugin %s', 'wpupstream' ), '<strong>WP Upstream</strong>' ); ?></p>
 		<p><?php printf( __( 'The plugin requires WordPress version %1$s. However, you are using version %2$s.', 'wpupstream' ), WP_UPSTREAM_REQUIRED_WP, $wp_version ); ?></p>
-		<p><?php _e( 'The plugin has been deactivated for now.', 'wpupstream' ); ?></p>
 	</div>
 	<?php
 }
@@ -25,7 +24,6 @@ function wpupstream_display_phpversion_error_notice() {
 	<div class="error">
 		<p><?php printf( __( 'Fatal error with plugin %s', 'wpupstream' ), '<strong>WP Upstream</strong>' ); ?></p>
 		<p><?php printf( __( 'The plugin requires PHP version %1$s. However, you are using version %2$s.', 'wpupstream' ), WP_UPSTREAM_REQUIRED_PHP, phpversion() ); ?></p>
-		<p><?php _e( 'The plugin has been deactivated for now.', 'wpupstream' ); ?></p>
 	</div>
 	<?php
 }
@@ -35,7 +33,6 @@ function wpupstream_display_spl_error_notice() {
 	<div class="error">
 		<p><?php printf( __( 'Fatal error with plugin %s', 'wpupstream' ), '<strong>WP Upstream</strong>' ); ?></p>
 		<p><?php _e( 'The PHP SPL functions can not be found. Please ask your hosting provider to enable them.', 'wpupstream' ); ?></p>
-		<p><?php _e( 'The plugin has been deactivated for now.', 'wpupstream' ); ?></p>
 	</div>
 	<?php
 }
@@ -50,12 +47,13 @@ function wpupstream_display_git_warning_notice() {
 }
 
 function wpupstream_deactivate() {
-	deactivate_plugins( WP_UPSTREAM_BASENAME );
-	if ( isset( $_GET['activate'] ) ) {
-		unset( $_GET['activate'] );
+	if ( is_admin() && current_user_can( 'activate_plugins' ) ) {
+		deactivate_plugins( WP_UPSTREAM_BASENAME );
+		if ( isset( $_GET['activate'] ) ) {
+			unset( $_GET['activate'] );
+		}
 	}
 }
-
 
 /**
  * Show indicator if enabled but broken or whatever else is critical.
@@ -63,16 +61,22 @@ function wpupstream_deactivate() {
  * @param $wp_admin_bar
  */
 function wpupstream_admin_bar_menu( $wp_admin_bar ) {
+	if ( ! is_super_admin() || ! $wp_admin_bar ) {
+		return;
+	}
 
-  if( !is_super_admin() || !is_multisite() || !$wp_admin_bar ) {
-    return;
-  }
+	$status = __( '%s is not running.', 'wpupstream' );
+	$color = '#dd3d36';
+	if ( class_exists( 'WPUpstream\Plugin' ) && WPUpstream\Plugin::has_instance() && WPUpstream\Plugin::instance()->get_status() ) {
+		$status = __( '%s is running.', 'wpupstream' );
+		$color = '#eeeeee';
+		//$color = '#7ad03a';
+	}
 
-  $wp_admin_bar->add_menu( array(
-    'id' => 'wp-upstream',
-    'parent' => 'top-secondary',
-    'title' => sprintf( __( '%s' ), '[]' ),
-    'href' => '#'
-  ));
-
+	$wp_admin_bar->add_menu( array(
+		'id'		=> 'wp-upstream',
+		'parent'	=> 'top-secondary',
+		'title'		=> '<span style="color:' . $color . ';">' . sprintf( $status, 'WP Upstream' ) . '</span>',
+		'href'		=> '#'
+	));
 }