+++ wp-upstream.php
@@ -1,22 +1,22 @@
 <?php
-/**
- * Plugin Name: WP-Upstream
- * Plugin URI: http://wordpress.org/plugins/wp-upstream/
- * Description: This plugin handles Git automation in WordPress.
- * Version: 0.1.1
- * Author: Usability Dynamics Inc.
- * Author URI: http://www.usabilitydynamics.com/
- * License: GNU General Public License v2
- * License URI: http://www.gnu.org/licenses/gpl-2.0.html
- * Text Domain: wpupstream
- * Domain Path: /languages/
- * Tags: wordpress, plugin, git, automization
- * GitHub Plugin URI: wpCloud/wp-upstream
- * GitHub Branch: v0.1
- * Network: True
+/*
+Plugin Name: WP Upstream
+Plugin URI: http://wordpress.org/plugins/wp-upstream/
+Description: This plugin handles Git automation in WordPress.
+Version: 0.1.2
+Author: Usability Dynamics Inc.
+Author URI: http://www.usabilitydynamics.com/
+License: GNU General Public License v2
+License URI: http://www.gnu.org/licenses/gpl-2.0.html
+Text Domain: wpupstream
+Domain Path: /languages/
+Tags: wordpress, plugin, git, automization
+GitHub Plugin URI: wpCloud/wp-upstream
+GitHub Branch: v0.1
+Network: True
 */
 /**
- * @package WPOD
+ * @package WPUpstream
  * @version 0.1.2
  * @author Usability Dynamics Inc.
  */
@@ -25,7 +25,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	die();
 }
 
-define( 'WP_UPSTREAM_VERSION', '0.1.1' );
+define( 'WP_UPSTREAM_VERSION', '0.1.2' );
 define( 'WP_UPSTREAM_REQUIRED_PHP', '5.3.0' );
 define( 'WP_UPSTREAM_REQUIRED_WP', '4.0' );
 
@@ -38,11 +38,9 @@ if ( ! defined( 'WP_UPSTREAM_DEBUG' ) ) {
 }
 
 function wpupstream() {
-
-  if( class_exists( 'WPUpstream\Plugin' ) ) {
-    return WPUpstream\Plugin::instance();
-  }
-
+	if ( class_exists( 'WPUpstream\Plugin' ) ) {
+		return WPUpstream\Plugin::instance();
+	}
 }
 
 function wpupstream_maybe_init() {
@@ -54,18 +52,14 @@ function wpupstream_maybe_init() {
 
 	add_action( 'plugins_loaded', 'wpupstream_load_textdomain', 1 );
 
-  // Admin bar status indicator.
-  add_action( 'admin_bar_menu', 'wpupstream_admin_bar_menu', 10 );
+  	// Admin bar status indicator.
+  	add_action( 'admin_bar_menu', 'wpupstream_admin_bar_menu', 10 );
 
-  if ( function_exists( 'spl_autoload_register' ) ) {
+  	if ( function_exists( 'spl_autoload_register' ) ) {
 
-    if ( file_exists( WP_UPSTREAM_PATH . '/vendor/autoload_52.php' ) ) {
-      require_once WP_UPSTREAM_PATH . '/vendor/autoload_52.php';
-		}
-
-    // Check for native autoload file.
-    if ( file_exists( WP_UPSTREAM_PATH . '/vendor/autoload.php' ) ) {
-      require_once WP_UPSTREAM_PATH . '/vendor/autoload.php';
+    	// Check for native autoload file.
+    	if ( file_exists( WP_UPSTREAM_PATH . '/vendor/autoload.php' ) ) {
+      		require_once WP_UPSTREAM_PATH . '/vendor/autoload.php';
 		}
 
 		if ( version_compare( phpversion(), WP_UPSTREAM_REQUIRED_PHP ) >= 0 ) {
@@ -80,10 +74,6 @@ function wpupstream_maybe_init() {
 		}
 	} else {
 		add_action( 'admin_notices', 'wpupstream_display_spl_error_notice' );
-	}
-
-	if ( ! $running && is_admin() ) {
-		add_action( 'admin_init', 'wpupstream_deactivate' );
 	}
 }