+++ changes.md
@@ -1,3 +1,13 @@
 #### 0.1.2
 * Changed all constant names to follow WP_UPSTREAM prefix.
 * Added changes.md file.
+* Added admin bar status notice.
+* Made plugin network only.
+* Removed PHP 5.2 autoloader.
+
+#### 0.1.1
+* fixed a bug where plugin would stop working after an unfinished process
+* added documentation
+
+#### 0.1.0
+* initial release